new changes

+++ FMS_Collection.Application/Services/CollectionCoinService.cs
@@ -11,10 +11,12 @@ namespace FMS_Collection.Application.Services
     {
         private readonly ICoinNoteCollectionRepository _repository;
         private readonly AzureBlobService _blobService;
-        public CoinNoteCollectionService(ICoinNoteCollectionRepository repository, AzureBlobService blobService)
+        private readonly AssetService _assetService;
+        public CoinNoteCollectionService(ICoinNoteCollectionRepository repository, AzureBlobService blobService, AssetService assetService)
         {
             _repository = repository;
             _blobService = blobService;
+            _assetService = assetService;
         }
 
         public async Task<ServiceResponse<List<CoinNoteCollection>>> GetAllCoinNoteCollectionsAsync()
@@ -37,18 +39,18 @@ namespace FMS_Collection.Application.Services
                 return response;
 
             // Replace ImagePath and ThumbnailPath with Blob SAS URLs
-            //foreach (var item in response.Data)
-            //{
-            //    if (!string.IsNullOrEmpty(item.ImagePath))
-            //    {
-            //        item.ImagePath = _blobService.GetBlobSasUrl(item.ImagePath);
-            //    }
+            foreach (var item in response.Data)
+            {
+                if (!string.IsNullOrEmpty(item.ImagePath))
+                {
+                    item.ImagePath = _blobService.GetBlobSasUrl(item.ImagePath);
+                }
 
-            //    if (!string.IsNullOrEmpty(item.ThumbnailPath))
-            //    {
-            //        item.ThumbnailPath = _blobService.GetBlobSasUrl(item.ThumbnailPath);
-            //    }
-            //}
+                if (!string.IsNullOrEmpty(item.ThumbnailPath))
+                {
+                    item.ThumbnailPath = _blobService.GetBlobSasUrl(item.ThumbnailPath);
+                }
+            }
 
             return response;
         }
@@ -61,7 +63,7 @@ namespace FMS_Collection.Application.Services
             );
         }
 
-        public async Task<ServiceResponse<Guid>> AddCoinNoteCollectionAsync(CoinNoteCollectionRequest CoinNoteCollection,Guid coinNoteCollectionId)
+        public async Task<ServiceResponse<Guid>> AddCoinNoteCollectionAsync(CoinNoteCollectionRequest CoinNoteCollection, Guid coinNoteCollectionId)
         {
             return await ServiceExecutor.ExecuteAsync(
                 () => _repository.AddAsync(CoinNoteCollection, coinNoteCollectionId),
@@ -79,6 +81,16 @@ namespace FMS_Collection.Application.Services
 
         public async Task<ServiceResponse<bool>> DeleteCoinNoteCollectionAsync(Guid coinNoteCollectionId, Guid userId)
         {
+            // first delete assets related to the selected coin/note
+            CoinNoteCollectionDetailsResponse coinDetails = await _repository.GetCoinNoteCollectionDetailsAsync(coinNoteCollectionId, userId);
+            var response = await _assetService.DeleteAssetAsync(coinDetails.AssetId, userId);
+            if (response == null)
+            {
+                return await ServiceExecutor.ExecuteAsync(
+                () => null,
+                FMS_Collection.Core.Constants.Constants.Messages.IssueInCoinDeletionNoteCollection
+            );
+            }
             return await ServiceExecutor.ExecuteAsync(
                 () => _repository.DeleteAsync(coinNoteCollectionId, userId),
                 FMS_Collection.Core.Constants.Constants.Messages.CoinNoteCollectionDeletedSuccessfully
